requirements.txt
flask
wakeonlan
speedtest-cli
hurry.filesize
requests
dload
psutils
